chore: move PaymentAwareComponent to a standalone file
